requirements.txt
# Core Testing Framework
pytest==8.3.3
pytest-playwright==0.5.2
playwright==1.48.0
pytest-xdist==3.6.1
pytest-rerunfailures==14.0

# Reporting
allure-pytest==2.13.5
pytest-html==4.1.1

# Configuration & Environment
python-dotenv==1.0.1
pyyaml==6.0.2

# Data Handling
faker==30.3.0
openpyxl==3.1.5
pandas==2.2.3

# Utilities
requests==2.32.3
pillow==11.0.0

# Type Checking & Linting
mypy==1.13.0
pylint==3.3.1
black==24.10.0
